use workflow to run spelling::spell_check_package()

This workflow will run the spell_check_package() function from the spelling package on the current package, and will fail if any spelling errors are found. It will also run the spell_check_files() on specified file types if spell_check_additional_files is TRUE.

use_spell_check(
  workflow_name = "call-spell-check.yml",
  build_trigger = c("pull_request", "push_to_main", "push_to_all_branches", "weekly",
    "manually"),
  spell_check_additional_files = FALSE,
  spell_check_report_level = c("error", "warning"),
  tag_ghactions4r = NULL
)

Arguments

workflow_name

What to name the github action workflow locally. Must have the file extension .yml

build_trigger

Select the build trigger. Options may include (depending on the workflow) running on pushing commits to main ("push_to_main"); run when a pull request is opened, reopened, or updated ("pull_request"); run when a comment containing the command \doc-and-style is made on a pull request by people with write permissions on the repository ("pr_comment"); run manually with the workflow_dispatch trigger ("manually"); run on the default branch (usually main) once a week ("weekly"). Currently, it is not possible to create a file with multiple build triggers from the function. Instead, users should add all of the appropriate build triggers by editing the file after the use_* function.

spell_check_additional_files

Logical. Should the workflow run spell_check_files() on specified files in the package? Defaults to FALSE.

spell_check_report_level

Character. The level of the report to generate for the additional spell check. Options are "warning" or "error". Defaults to "error". If set to "error", the workflow will fail if any spelling errors are found. If set to "warning", the the workflow will pass even if spelling errors are found.

tag_ghactions4r

Tag of ghactions4r workflow to pin to for the reusable workflow. This ensures compatibility with a specific version of ghactions4r functions. The tradeoff is that fixes will not happen automatically; instead, the user will have to update the pinned version of the caller workflow used. See the options by looking at the ghactions4r tags. Another option is to pin to commit hashes, but the user will need to do this by manually editing the caller workflow after creation. See github documentation on security to understand how this is more secure. Again, the tradeoff is that the workflow will not be automatically updated.